clean up weather forecasting project
