--- projects/weather_forecasting/flytelab/weather_forecasting/v1/__init__.py
@@ -1,5 +1,0 @@
-"""
-NOTE: This version of the weather forecasting workflow is no longer maintained
-"""
-
-from flytelab.weather_forecasting.v1 import data, trainer
